backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "pyjwt>=2.8.0",
    "werkzeug>=3.0.0",
    "requests>=2.31.0",
    "cryptography>=41.0.0",
    "openai>=1.0.0",
    "mem0ai[graph]>=1.0.1",
    "python-dotenv>=1.2.1",
    "google-generativeai>=0.8.5",
    "google-genai>=1.54.0",
    "sentence-transformers>=5.2.2",
]
